refactor(pr06): rename IdentifiableItem.mts -> IdentifiableDocument.mts (G5b)
File-only rename to match class name (class is IdentifiableDocument). Local variable names like IdentifiableItemBase intentionally retained per spec.
